Extract Variable to separate module
Before introducing the new `VariableRefMut` type, this commit extracts
the `Variable` type to a separate module for better cohesion.
